+++ include/ProfileInfoMerge.h
@@ -16,7 +16,7 @@ class ProfileInfoMerge
    std::vector<std::string> CommandLines;
    std::vector<unsigned> FunctionCounts;
    std::vector<uint64_t> BlockCounts;
-   std::vector<unsigned> EdgeCounts;
+   std::vector<uint64_t> EdgeCounts;
    std::vector<unsigned> OptimalEdgeCounts;
    std::vector<unsigned> BBTrace;
    std::vector<unsigned> ValueCounts;
